requirements.txt
openai>=1.99.0
pydantic>=2.9
python-dotenv>=1.0
rich>=13.7
typer>=0.12
